+++ client/flags/flags.go
@@ -32,10 +32,6 @@ const (
 	SignModeDirect = "direct"
 	// SignModeLegacyAminoJSON is the value of the --sign-mode flag for SIGN_MODE_LEGACY_AMINO_JSON
 	SignModeLegacyAminoJSON = "amino-json"
-	// SignModeDirectAux is the value of the --sign-mode flag for SIGN_MODE_DIRECT_AUX
-	SignModeDirectAux = "direct-aux"
-	// SignModeTextual is the value of the --sign-mode flag for SIGN_MODE_TEXTUAL.
-	SignModeTextual = "textual"
 	// SignModeEIP191 is the value of the --sign-mode flag for SIGN_MODE_EIP_191
 	SignModeEIP191 = "eip-191"
 )